% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analyze_data_files}
\alias{analyze_data_files}
\title{Analyze each data set in a folder}
\usage{
analyze_data_files(path_in = "./")
}
\arguments{
\item{path_in}{A character vector. Path to data folder}
}
\description{
Analyze each data set in a folder
}
\examples{
library(rWorkshop1Demo)
analyze_data_files(path_in = './data/')
list.files('./output/')
}

