% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean_data_files}
\alias{clean_data_files}
\title{Clean all data files in a folder}
\usage{
clean_data_files(path_in = "./", path_out = "./")
}
\arguments{
\item{path_in}{A character vector. Path to data folder}

\item{path_out}{A character vector. Path to output folder: Where cleaned data files will be saved as .csv}
}
\description{
Clean all data files in a folder
}
\examples{
library(rWorkshop1Demo)
clean_data_files(path_in = './data/', path_out = './output/')
list.files('./output/')
}

