% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_data}
\alias{create_data}
\title{Generate random toy datasets for each US state}
\usage{
create_data(size = 1000, seed = NULL, path = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{key,value}{Names of key and value columns to create in output.}

\item{Specification}{of columns to gather. Use bare variable names.}
}
\description{
Generate random toy datasets for each US state
}
\examples{
library(rWorkshop1Demo)
create_data()
list.files()
}

