% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_data_files}
\alias{plot_data_files}
\title{Create a plot for each data set in a folder}
\usage{
plot_data_files(path_in = "./", path_out = "./")
}
\arguments{
\item{path_in}{A character vector. Path to data folder}

\item{path_out}{A character vector. Path to output folder: Where plots will be saved as .png}
}
\description{
Create a plot for each data set in a folder
}
\examples{
library(rWorkshop1Demo)
plot_data_files(path_in = './data/', path_out = './output/')
list.files('./output/')
}

